Refactor app from folder-by-tech to folder-by-feature
