+++ src/common/endpoint.interface.ts
