+++ src/common/types/redux.ts
@@ -1,5 +1,5 @@
-import { RootAction } from "@src/actions/interfaces";
-import { RootState } from "@src/state/state";
+import { RootAction } from "@src/redux/interfaces";
+import { RootState } from "@src/redux/state";
 
 export type Reducer<State> = (state: State|undefined, action: Readonly<RootAction>) => State;
