+++ src/features/comments/comment-data.interface.ts
@@ -1,0 +1,5 @@
+export interface CommentData {
+    id: string;
+    username: string;
+    text: string;
+}
