+++ src/features/photos/photo-data.interface.ts
@@ -1,0 +1,11 @@
+import { CommentData } from "@src/features/comments/comment-data.interface";
+
+export interface PhotoData {
+    id: string;
+    link: string;
+    comments: CommentData[];
+    images: {
+        standard: string;
+        thumbnail: string;
+    };
+}
