+++ src/redux/helpers.ts
@@ -1,5 +1,5 @@
-import { ThunkAction } from "@src/types/redux";
-import { RootAction } from "@src/actions/interfaces";
+import { ThunkAction } from "@src/common/types/redux";
+import { RootAction } from "@src/redux/interfaces";
 
 // type ActionCreator<T extends DispatchFunction | ActionFuncion> = T;
 type ActionFuncion = (...args: any[]) => RootAction;
