+++ src/redux/state.ts
@@ -1,16 +1,14 @@
 import { RouterState } from "connected-react-router";
 
-import { Photo } from "@src/interfaces/data";
+import { PhotoData } from "@src/features/photos/photo-data.interface";
 
 export interface RootState {
     readonly router: RouterState;
-    // readonly selectedPhoto: SelectedPhotoState;
     readonly photos: PhotosState;
 }
 
-// export type SelectedPhotoState = Readonly<Photo>|null;
 export interface PhotosState {
     profileName: string;
-    photos: ReadonlyArray<Photo>;
+    photos: ReadonlyArray<PhotoData>;
     isFetching: boolean;
 }
