+++ src/redux/types.ts
@@ -1,5 +1,4 @@
 export enum ActionType {
-    // SELECT_PHOTO = "SELECT_PHOTO",
     PHOTOS_FETCHED = "PHOTOS/FETCHED",
     PHOTOS_FETCH_STARTED = "PHOTOS/FETCH_STARTED",
 }
